move the pendingInput and model client to related object
