+++ codex-chrome/src/core/session/state/__tests__/Persistence.integration.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect, beforeEach } from 'vitest';
-import { Session } from '../../Session';
+import { Session } from '../../../Session';
 import { SessionState } from '../SessionState';
 import type { InputItem } from '../../../protocol/types';
