+++ codex-chrome/src/core/session/state/__tests__/Session.integration.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { Session } from '../../Session';
+import { Session } from '../../../Session';
 import type { InputItem } from '../../../protocol/types';
 
 describe('Session Integration (Refactored)', () => {
