+++ codex-chrome/src/core/session/state/__tests__/TurnExecution.integration.test.ts
@@ -5,10 +5,10 @@
  */
 
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { Session } from '../../Session';
+import { Session } from '../../../Session';
 import { TaskKind } from '../types';
 import type { InputItem } from '../../../protocol/types';
-import { ReviewDecision } from '../../../protocol/types';
+import { ReviewDecision } from '../../../../protocol/types';
 
 describe('Turn Execution Integration', () => {
   let session: Session;
